Refactored Response classes
