+++ src/SaaS/Http/Response.php
@@ -25,21 +25,12 @@ use SaaS\Exception\InvalidJsonException;
  * @link     http://github.com/gwinn/saas-connector
  * @see      http://github.com/gwinn/saas-connector
  */
-class Response implements \ArrayAccess
+class Response extends BaseResponse
 {
-    protected $statusCode;
-    protected $response;
     protected $responseBody;
 
-    /**
-     * Response constructor.
-     *
-     * @param integer $statusCode   status code
-     * @param mixed   $responseBody response body content
-     */
-    public function __construct($statusCode, $responseBody = null)
+    protected function processResponse($responseBody)
     {
-        $this->statusCode = (int) $statusCode;
         $this->responseBody = $responseBody;
 
         if (!empty($responseBody)) {
@@ -76,135 +67,12 @@ class Response implements \ArrayAccess
     }
 
     /**
-     * Return HTTP response status code
-     *
-     * @return int
-     */
-    public function getStatusCode()
-    {
-        return $this->statusCode;
-    }
-
-    /**
-     * Return HTTP response body
-     *
-     * @return mixed
-     */
-    public function getResponse()
-    {
-        return $this->response;
-    }
-
-    /**
-     * Return HTTP response row
+     * Return HTTP response raw
      *
      * @return mixed
      */
     public function getResponseBody()
     {
         return $this->responseBody;
-    }
-
-    /**
-     * HTTP request was successful
-     *
-     * @return bool
-     */
-    public function isSuccessful()
-    {
-        return $this->statusCode < 400;
-    }
-
-    /**
-     * Allow to access for the property throw class method
-     *
-     * @param string $name      method name
-     * @param array  $arguments method arguments
-     *
-     * @return mixed
-     */
-    public function __call($name, $arguments)
-    {
-        // convert getSomeProperty to someProperty
-        $propertyName = strtolower(substr($name, 3, 1)) . substr($name, 4);
-
-        if (!isset($this->response[$propertyName])) {
-            throw new \InvalidArgumentException("Method \"$name\" not found");
-        }
-
-        return $this->response[$propertyName];
-    }
-
-    /**
-     * Allow to access for the property throw object property
-     *
-     * @param string $name property name
-     *
-     * @return mixed
-     */
-    public function __get($name)
-    {
-        if (!isset($this->response[$name])) {
-            throw new \InvalidArgumentException("Property \"$name\" not found");
-        }
-
-        return $this->response[$name];
-    }
-
-    /**
-     * Set offset
-     *
-     * @param mixed $offset offset value
-     * @param mixed $value  value
-     *
-     * @return void
-     */
-    public function offsetSet($offset, $value)
-    {
-        throw new \BadMethodCallException(
-            "This call not allowed: offsetSet [$offset] [$value]"
-        );
-    }
-
-    /**
-     * Unset offset
-     *
-     * @param mixed $offset offset value
-     *
-     * @return void
-     */
-    public function offsetUnset($offset)
-    {
-        throw new \BadMethodCallException(
-            "This call not allowed: offsetSet [$offset]"
-        );
-    }
-
-    /**
-     * Check offset
-     *
-     * @param mixed $offset offset value
-     *
-     * @return bool
-     */
-    public function offsetExists($offset)
-    {
-        return isset($this->response[$offset]);
-    }
-
-    /**
-     * Get offset
-     *
-     * @param mixed $offset offset value
-     *
-     * @return mixed
-     */
-    public function offsetGet($offset)
-    {
-        if (!isset($this->response[$offset])) {
-            throw new \InvalidArgumentException("Property \"$offset\" not found");
-        }
-
-        return $this->response[$offset];
     }
 }
